# hugolib, tpl: Get rid of some redundant viper calls