@bep 2017-04-06
In my tests it saves some allocations compared to current master, but is about the same in speed department (there is an error margin here at about 3 percent or something):

```
benchmark           old ns/op      new ns/op       delta
BenchmarkHugo-4     9997747579     10140845478     +1.43%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     49446752       48760618       -1.39%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9442790920     9433569344     -0.10%
```