@bogem 2017-04-06
Hm
Then these changes only affects themes that uses many `markdownify` functions