@bep 2017-04-06
This is, of course, still worth it. I will eventually, update my list of sites in my benchmark to be more representative re new features.